Revert "Feature#14"

+++ weather/src/main/java/com/practice/weather/shortTerm/extra/repository/ShortTermExtraRepository.java
@@ -19,6 +19,6 @@ public interface ShortTermExtraRepository extends JpaRepository<ShortTermExtraEn
     List<ShortTermExtraEntity> selectListByLocation(Pageable pageable, String nxValue, String nyValue);
 
     @Query(value = "SELECT COUNT(1) FROM SHORT_TERM_EXTRA WHERE NX_VALUE = :nxValue AND NY_VALUE = :nyValue", nativeQuery = true)
-    long countByLocation(String nxValue, String nyValue);
+    int countByLocation(String nxValue, String nyValue);
     
 }
